refactor socket.io with egg-socket.io
