+++ app/io/controller/tomatobang.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = () => {
+  return async function () {
+      // 这里暂时不做处理，交由中间件处理
+  };
+};
