+++ config/plugin.js
@@ -19,3 +19,8 @@ exports.cors = {
   enable: true,
   package: 'egg-cors',
 };
+
+exports.io = {
+  enable: true,
+  package: 'egg-socket.io',
+};
